// src/types/AnimeProps.ts
export type AnimeProps = {
  universal_anime_id: string;
  integration_service: string;
  integration_id: string;
  title: string;
  alternative_name: string;
  sub: string;
  resume: string;
  rateing: string;
  image: string;
  created_at: string;
  updated_at: string;
  status: string;
  status_describe: string;
  attempts_to_cancel_updates: string;
};